+++ src/api/auth/login.rs
@@ -7,7 +7,7 @@ use crate::errors::ChorusResult;
 use crate::gateway::Gateway;
 use crate::instance::{ChorusUser, Instance};
 use crate::ratelimiter::ChorusRequest;
-use crate::types::{GatewayIdentifyPayload, LimitType, LoginResult, LoginSchema, User};
+use crate::types::{GatewayIdentifyPayload, LimitType, LoginResult, LoginSchema};
 
 impl Instance {
     /// Logs into an existing account on the spacebar server.
@@ -45,31 +45,6 @@ impl Instance {
             self.clone_limits_if_some(),
             login_result.settings,
             Arc::new(RwLock::new(object)),
-            gateway,
-        );
-        Ok(user)
-    }
-
-    /// Logs into an existing account on the spacebar server, using only a token.
-    pub async fn login_account_with_token(&mut self, token: String) -> ChorusResult<ChorusUser> {
-        let object_result = self.get_user(token.clone(), None).await;
-        if let Err(e) = object_result {
-            return Result::Err(e);
-        }
-
-        let user_settings = User::get_settings(&token, &self.urls.api, &mut self.clone())
-            .await
-            .unwrap();
-        let mut identify = GatewayIdentifyPayload::common();
-        let gateway = Gateway::new(self.urls.wss.clone()).await.unwrap();
-        identify.token = token.clone();
-        gateway.send_identify(identify).await;
-        let user = ChorusUser::new(
-            Arc::new(RwLock::new(self.clone())),
-            token.clone(),
-            self.clone_limits_if_some(),
-            Arc::new(RwLock::new(user_settings)),
-            Arc::new(RwLock::new(object_result.unwrap())),
             gateway,
         );
         Ok(user)
